Update README.md
Solving N-Queen Problem by Prediction :
The N-Queen problem is a famous challenge in computer science. The task is to place N queens on an N×N chessboard so that no two queens attack each other. This means no two queens can share the same row, column, or diagonal.Traditionally, backtracking is used to solve it. While effective, it can be slow for larger N due to high computational complexity.
We’ll look at a smarter, faster method: solving the N-Queen problem by prediction a novel algorithm that reduces the number of unnecessary trials.